+++ ecmascript/builtins/builtins_typedarray.cpp
@@ -182,7 +182,7 @@ JSTaggedValue BuiltinsTypedArray::From(EcmaRuntimeCallInfo *argv)
                 vec.push_back(nextValue);
             }
         }
-        int32_t len = vec.size();
+        uint32_t len = vec.size();
         JSTaggedType args[1] = {JSTaggedValue(len).GetRawData()};
         JSHandle<JSObject> targetObj = TypedArrayHelper::TypedArrayCreate(thread, thisHandle, 1, args);
         RETURN_EXCEPTION_IF_ABRUPT_COMPLETION(thread);
@@ -775,13 +775,14 @@ JSTaggedValue BuiltinsTypedArray::Join(EcmaRuntimeCallInfo *argv)
             if (sep >= 0) {
                 newString->WriteData(static_cast<char>(sep), current);
             } else if (sep != BuiltinsTypedArray::SeparatorFlag::MINUS_TWO) {
-                newString->WriteData(*sepStringHandle, current, allocateLength - current, sepLength);
+                newString->WriteData(
+                    *sepStringHandle, current, allocateLength - static_cast<size_t>(current), sepLength);
             }
-            current += sepLength;
+            current += static_cast<int>(sepLength);
         }
         JSHandle<EcmaString> nextStr = vec[k];
         int nextLength = static_cast<int>(nextStr->GetLength());
-        newString->WriteData(*nextStr, current, static_cast<size_t>(allocateLength - current), nextLength);
+        newString->WriteData(*nextStr, current, allocateLength - static_cast<size_t>(current), nextLength);
         current += nextLength;
     }
     return JSTaggedValue(newString);
